Refactored font_table.c:font_table_create() and the associated create functions to take in a buffer of bytes instead of a file pointer. Checksum verification moved out of each create function.

+++ src/ttf_loader/tables/os2.h
@@ -41,6 +41,6 @@ typedef struct os2_table {
   uint16_t us_upper_optical_point_size;
 } os2_table_t;
 
-os2_table_t* os2_table_create(uint32_t* table_data, table_directory_t* table_directory);
+os2_table_t* os2_table_create(uint8_t* table_data, table_directory_t* table_directory);
 
 #endif // !DEBUG
